Add an attr check before calling type()
When a value within the dict passed to urlencode() is not a string, the
check `v.type(str)` will throw an AttributeException. This is handled
correctly further down in the urlencode function, but we need to check
that the value actually has a `type` attribute before trying to call
`type`.